Fix zero variance in a Normalization layer (#384)
If one or more of the variance values in a Normalization layer is zero, for example if the classifier was trained on images that always have a black pixel in the top-left corner, a division by zero will cause all model responses to be NaN. Keras handles this by replacing zero values with an epsilon value that defaults to 1e-7. This library should do the same - preferably with a parametric epsilon but at least with the same epsilon as Keras.